+++ EveryCert/EveryCert/src/Model/Element/ElementHandler.m
@@ -220,7 +220,7 @@
 
 - (NSString *)getApiCallWithFormId:(NSInteger)formId
 {
-    return [NSString stringWithFormat:@"%@/%ld",self.apiName, formId];
+    return [NSString stringWithFormat:@"%@/%ld",self.apiName, (long)formId];
 }
 
 - (void)syncWithServer
