+++ EveryCert/EveryCert/src/Model/SubElement/SubElementHandler.m
@@ -89,7 +89,7 @@
 
 - (NSString *)getApiCallWithFormId:(NSInteger)formId
 {
-    return [NSString stringWithFormat:@"%@/%ld",self.apiName, formId];
+    return [NSString stringWithFormat:@"%@/%ld",self.apiName, (long)formId];
 }
 
 - (void)syncWithServer
